+++ exercise/src/main/java/com/bidding/prediction/domain/persistence/CoefficientRepository.java
@@ -1,10 +1,11 @@
 package com.bidding.prediction.domain.persistence;
 
+import java.math.BigDecimal;
 import java.util.Map;
 import java.util.Set;
 
 public interface CoefficientRepository {
 
-	Map<String, Double> getCoefficients(Set<String> featureNames);
-	
+	Map<String, BigDecimal> getCoefficients(Set<String> featureNames);
+
 }
